debundle/lowering: extract lower
BAZEL_TEST_INVOCATIONS=none: file move.
